Move big tests into their own files.
